requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
sentence-transformers
transformers
langchain
flask
pypdf
python-dotenv
pinecone-client
pinecone[grpc]
langchain-pinecone
huggingface-hub
gunicorn
langchain_community
langchain-huggingface 